app/requirements.txt
streamlit==1.32.0
groq
spacy
pandas==2.2.1
plotly==5.20.0
python-dotenv==1.0.1
pydantic==2.6.4
